[qwen3_5] add qwen3_5 dense and moe model
1. rename Qwen3_vl_next to Qwen3_5 in config
2. merge qwen3_5.py and qwen3_5_moe.py
