+++ python/sglang/srt/configs/__init__.py
@@ -19,7 +19,7 @@ from sglang.srt.configs.nano_nemotron_vl import NemotronH_Nano_VL_V2_Config
 from sglang.srt.configs.nemotron_h import NemotronHConfig
 from sglang.srt.configs.olmo3 import Olmo3Config
 from sglang.srt.configs.qwen3_next import Qwen3NextConfig
-from sglang.srt.configs.qwen3_next_vl import Qwen3NextVLConfig, Qwen3NextVLMoEConfig
+from sglang.srt.configs.qwen3_5 import Qwen3_5Config, Qwen3_5MoeConfig
 from sglang.srt.configs.step3_vl import (
     Step3TextConfig,
     Step3VisionEncoderConfig,
@@ -44,8 +44,8 @@ __all__ = [
     "KimiLinearConfig",
     "KimiK25Config",
     "Qwen3NextConfig",
-    "Qwen3NextVLConfig",
-    "Qwen3NextVLMoEConfig",
+    "Qwen3_5Config",
+    "Qwen3_5MoeConfig",
     "DotsVLMConfig",
     "DotsOCRConfig",
     "FalconH1Config",
