+++ python/sglang/srt/model_executor/model_runner.py
@@ -39,8 +39,8 @@ from sglang.srt.configs import (
     NemotronH_Nano_VL_V2_Config,
     NemotronHConfig,
     Qwen3NextConfig,
-    Qwen3NextVLConfig,
-    Qwen3NextVLMoEConfig,
+    Qwen3_5Config,
+    Qwen3_5MoeConfig,
 )
 from sglang.srt.configs.device_config import DeviceConfig
 from sglang.srt.configs.load_config import LoadConfig, LoadFormat
@@ -1501,7 +1501,7 @@ class ModelRunner(ModelRunnerKVCacheMixin):
     @property
     def hybrid_gdn_config(self):
         config = self.model_config.hf_config.get_text_config()
-        if isinstance(config, Qwen3NextConfig | Qwen3NextVLConfig | Qwen3NextVLMoEConfig | JetNemotronConfig | JetVLMConfig):
+        if isinstance(config, Qwen3NextConfig | Qwen3_5Config | Qwen3_5MoeConfig | JetNemotronConfig | JetVLMConfig):
             return config
         return None
