+++ python/sglang/srt/multimodal/processors/qwen_vl.py
@@ -18,8 +18,7 @@ from sglang.srt.models.qwen2_vl import Qwen2VLForConditionalGeneration
 from sglang.srt.models.qwen3_omni_moe import Qwen3OmniMoeForConditionalGeneration
 from sglang.srt.models.qwen3_vl import Qwen3VLForConditionalGeneration
 from sglang.srt.models.qwen3_vl_moe import Qwen3VLMoeForConditionalGeneration
-from sglang.srt.models.qwen3_5 import Qwen3_5ForConditionalGeneration
-from sglang.srt.models.qwen3_5_moe import Qwen3_5MoeForConditionalGeneration
+from sglang.srt.models.qwen3_5 import Qwen3_5ForConditionalGeneration, Qwen3_5MoeForConditionalGeneration
 
 from sglang.srt.multimodal.processors.base_processor import (
     BaseMultimodalProcessor as SGLangBaseProcessor,
