+++ python/sglang/srt/utils/hf_transformers_utils.py
@@ -66,6 +66,8 @@ from sglang.srt.configs import (
     Step3p5Config,
     Qwen3NextVLConfig,
     Qwen3NextVLMoEConfig,
+    Qwen3_5Config,
+    Qwen3_5MoeConfig,
     Step3VLConfig,
 )
 from sglang.srt.configs.deepseek_ocr import DeepseekVLV2Config
@@ -95,8 +97,8 @@ _CONFIG_REGISTRY: List[Type[PretrainedConfig]] = [
     NemotronH_Nano_VL_V2_Config,
     NemotronHConfig,
     DeepseekVLV2Config,
-    Qwen3NextVLConfig,
-    Qwen3NextVLMoEConfig,
+    Qwen3_5Config,
+    Qwen3_5MoeConfig,
     JetNemotronConfig,
     JetVLMConfig,
     KimiK25Config,
